Add the ability to differential the display of certain types of credentials.
